+++ release_notes.md
@@ -10,4 +10,4 @@
   - Replaced `InvalidOperationException` with `HttpForwardingException` when there is a ForwarderError
 - [In-proc] Codeql : Fix to remove exception details from the response (#10751)
 - Update PowerShell 7.4 worker to [4.0.4134](https://github.com/Azure/azure-functions-powershell-worker/releases/tag/v4.0.4134)
-- Add support for new appsetting `FUNCTIONS_AZUREMONITOR_TIME_ISOFORMAT_ENABLED` to enable iso time format for azmon logs for Linux Dedicated/EP Skus. (part of #7864)
+- Add support for new FeatureFlag `EnableAzureMonitorTimeIsoFormat` to enable iso time format for azmon logs for Linux Dedicated/EP Skus. (part of #7864)